+++ src/widget/toggle_button_widget.rs
@@ -38,7 +38,6 @@ pub struct ToggleButtonWidget {
     active: bool,
     widget_id: i32,
     callbacks: DefaultWidgetCallbacks,
-    on_click: Option<Box<dyn FnMut(&mut dyn Widget, bool, &Vec<WidgetContainer>)>>,
 }
 
 impl ToggleButtonWidget {
@@ -60,7 +59,6 @@ impl ToggleButtonWidget {
             active: false,
             widget_id: 0,
             callbacks: DefaultWidgetCallbacks::new(),
-            on_click: None,
         }
     }
 
@@ -92,23 +90,14 @@ impl ToggleButtonWidget {
         self.invalidate();
     }
 
-    /// Sets a callback closure that can be called when a click is registered for this
-    /// widget.
-    pub fn on_click<F>(&mut self, callback: F)
-    where
-        F: FnMut(&mut dyn Widget, bool, &Vec<WidgetContainer>) + 'static,
-    {
-        self.on_click = Some(Box::new(callback));
-    }
-
     /// Calls the click `on_click` callback, if set.  Otherwise, ignored.  Sends a reference
     /// of the current `Widget` object as a parameter, so this object can be modified when
     /// a click is registered, if necessary.  Also indicates the state of the object, whether
     /// or not the object has been toggled.
     pub fn click(&mut self, state: bool, widgets: &Vec<WidgetContainer>) {
-        if let Some(mut cb) = self.on_click.take() {
+        if let Some(mut cb) = self.get_callbacks().on_toggle.take() {
             cb(self, state, widgets);
-            self.on_click = Some(cb);
+            self.get_callbacks().on_toggle = Some(cb);
         }
     }
 }