+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 - Moved callbacks to `DefaultWidgetCallbacks` object so that callbacks are stored in a sturcture
 - Converted callbacks to use get_callbacks() function for setting callback closures
 - Modified PushButtonWidget to use get_callbacks()
+- Added on_toggle callback (#190)
 
 ## 0.4.4